pyproject.toml
[tool.poetry]
name = "pdfmarks"
version = "0.1.0"
description = "pdfs from bookmarks"
authors = ["Sander van Dragt <sander@vandragt.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pdfkit = "^1.0.0"
python-slugify = "^6.1.2"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
